# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a microservice for managing tasks and users.",
        "title": "Task & User Microservice API",
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
        "/task": {
            "get": {
                "summary": "Get all tasks",
                "description": "Retrieve all tasks from the system",
                "produces": ["application/json"],
                "responses": {
                    "200": {
                        "description": "A list of tasks",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Task"
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a new task",
                "description": "Add a new task to the list",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "parameters": [
                    {
                        "in": "body",
                        "name": "task",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task created"
                    }
                }
            }
        },
        "/task/{id}": {
            "get": {
                "summary": "Get task by ID",
                "description": "Retrieve a specific task by its ID",
                "produces": ["application/json"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task object",
                        "schema": {
                            "$ref": "#/definitions/Task"
                        }
                    },
                    "404": {
                        "description": "Task not found"
                    }
                }
            },
            "patch": {
                "summary": "Complete task",
                "description": "Mark a task as complete by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task completed"
                    }
                }
            },
            "delete": {
                "summary": "Delete task",
                "description": "Delete task by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task deleted"
                    }
                }
            }
        }
    },
    "definitions": {
        "Task": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}

 